requirements.txt
mlflow==2.16.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
joblib==1.4.2
